Get information on a Member of the Scottish Parliament

Get information on a Member of the Scottish Parliament

getMSP(id = NULL, postcode = NULL, constituency = NULL,
  always_return = NULL)

Arguments

id
Person identifier

postcode
A postcode, which specifies the constiuency whose MSP is required

constituency
Name of a constituency

always_return
whether to try to return an MP even if the seat is vacant or it before an election

Value

a data.frame with rows representing the MP's spells in Parliament and columns

member_id Member identifier for each spell in parliament

house 1: House of Commons, 2: House of Lords, 3: Legislative Assembly (NI), 4: Scottish Parliament

constituency Constituency represented

party Party

entered_house Date MP entered Parliament

left_house Date MP left Parliament, or 9999-12-31 if still in place

entered_reason Reason MP entered, e.g. general_election

left_reason Reason MP left, e.g. general_election_standing

person_id Person identifier

lastupdate When TheyWorkForYou last updated this information

title Title, if any

given_name First name

family_name Family name

full_name First name and family name

url URL path relative to TheyWorkForYou's hostname

image URL path to jpg relative to TheyWorkForYou's hostname

image_height Image height in pixels

image_width Image width in pixels